correção em todas as entidades persistentes nos pacotes, model, service, repository. Problema resolvido: lazy initialization exception.

+++ bores/src/main/resources/db/migration/V01__categoria.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "categoria" (
   "id" serial8,
   "nome" varchar(30) NOT NULL,
-  "version" int8 NOT NULL,
+  "version" int4 NOT NULL,
   PRIMARY KEY ("id")
 );
