+++ bores/src/main/resources/db/migration/V02__subcategoria.sql
@@ -2,7 +2,7 @@ CREATE TABLE "subcategoria" (
   "id" serial8,
   "nome" varchar(30) NOT NULL,
   "categoria_id" int8 NOT NULL,
-  "version" int8 NOT NULL,
+  "version" int4 NOT NULL,
   PRIMARY KEY ("id")
 );
